+++ protected/humhub/modules/admin/messages/nb_no/forms_LogsSettingsForm.php
@@ -1,10 +1,10 @@
 <?php
 return array (
-  '1 month' => '',
-  '1 week' => '',
-  '1 year' => '',
-  '2 weeks' => '',
-  '3 months' => '',
-  '6 months' => '',
+  '1 month' => 'En måned',
+  '1 week' => 'En uke',
+  '1 year' => 'Ett år',
+  '2 weeks' => 'To Uker',
+  '3 months' => 'Tre måneder',
+  '6 months' => 'Ett halvt år',
   'never' => 'aldri',
 );