+++ protected/humhub/modules/admin/messages/nb_no/forms_MailingSettingsForm.php
@@ -3,9 +3,9 @@ return array (
   'Allow Self-Signed Certificates?' => '',
   'E-Mail sender address' => '',
   'E-Mail sender name' => '',
-  'Encryption' => '',
+  'Encryption' => 'Kryptering',
   'Mail Transport Type' => '',
   'Password' => 'Passord',
-  'Port number' => '',
+  'Port number' => 'Port nummer',
   'Username' => 'Brukernavn',
 );