+++ protected/humhub/modules/admin/messages/nb_no/views_approval_approveUserAccept.php
@@ -1,6 +1,6 @@
 <?php
 return array (
-  'Accept user: <strong>{displayName}</strong> ' => '',
+  'Accept user: <strong>{displayName}</strong> ' => 'Aksepter bruker: <strong>{displayName}</strong>',
   'Cancel' => 'Avbryt',
-  'Send & save' => '',
+  'Send & save' => 'Send & lager',
 );