+++ protected/humhub/modules/content/messages/ja/widgets_views_archiveLink.php
@@ -1,5 +1,5 @@
 <?php
 return array (
-  'Move to archive' => '',
+  'Move to archive' => 'アーカイブへ移動する',
   'Unarchive' => '解凍する',
 );