+++ protected/humhub/modules/content/models/Content.php
@@ -37,6 +37,8 @@ use yii\rbac\Permission;
  * @property string $updated_at
  * @property integer $updated_by
  * @property ContentContainer $contentContainer
+ * @property string $stream_sort_date
+ * @property string $stream_channel
  * @property integer $contentcontainer_id;
  * @property ContentContainerActiveRecord $container
  *
@@ -136,11 +138,9 @@ class Content extends ContentDeprecated
      */
     public function beforeSave($insert)
     {
-
         if ($this->object_model == "" || $this->object_id == "") {
             throw new Exception("Could not save content with object_model or object_id!");
         }
-
 
         // Set some default values
         if (!$this->archived) {
@@ -393,7 +393,7 @@ class Content extends ContentDeprecated
     {
         $this->contentcontainer_id = $container->contentContainerRecord->id;
         $this->_container = $container;
-        if($container instanceof Space && $this->visibility === null) {
+        if ($container instanceof Space && $this->visibility === null) {
             $this->visibility = $container->getDefaultContentVisibility();
         }
     }
@@ -460,11 +460,11 @@ class Content extends ContentDeprecated
      */
     public function addTag(ContentTag $tag)
     {
-        if(!empty($tag->contentcontainer_id) && $tag->contentcontainer_id != $this->contentcontainer_id) {
+        if (!empty($tag->contentcontainer_id) && $tag->contentcontainer_id != $this->contentcontainer_id) {
             throw new InvalidParamException(Yii::t('ContentModule.base', 'Content Tag with invalid contentcontainer_id assigned.'));
         }
 
-        if(ContentTagRelation::findBy($this, $tag)->count()) {
+        if (ContentTagRelation::findBy($this, $tag)->count()) {
             return true;
         }
 
@@ -473,7 +473,6 @@ class Content extends ContentDeprecated
         $contentRelation = new ContentTagRelation($this, $tag);
         return $contentRelation->save();
     }
-
 
     /**
      * Checks if the given user can edit this content.
@@ -555,6 +554,11 @@ class Content extends ContentDeprecated
             $user = Yii::$app->user->getIdentity();
         }
 
+        // User cann access own content
+        if ($user !== null && $this->created_by == $user->id) {
+            return true;
+        }
+
         // Check Guest Visibility
         if (!$user) {
             return $this->checkGuestAccess();
@@ -569,8 +573,8 @@ class Content extends ContentDeprecated
         if ($user->isSystemAdmin() && Yii::$app->getModule('content')->adminCanViewAllContent) {
             return true;
         }
-
-        if ($this->isPrivate() && $this->getContainer()->canAccessPrivateContent($user)) {
+        
+        if ($this->isPrivate() && $this->getContainer() !== null && $this->getContainer()->canAccessPrivateContent($user)) {
             return true;
         }
 
@@ -589,13 +593,12 @@ class Content extends ContentDeprecated
      */
     public function checkGuestAccess()
     {
-        if(!$this->isPublic() || !Yii::$app->getModule('user')->settings->get('auth.allowGuestAccess')) {
+        if (!$this->isPublic() || !Yii::$app->getModule('user')->settings->get('auth.allowGuestAccess')) {
             return false;
         }
 
         // Check container visibility for guests
-        return ($this->container instanceof Space && $this->container->visibility == Space::VISIBILITY_ALL)
-            || ($this->container instanceof User && $this->container->visibility == User::VISIBILITY_ALL);
+        return ($this->container instanceof Space && $this->container->visibility == Space::VISIBILITY_ALL) || ($this->container instanceof User && $this->container->visibility == User::VISIBILITY_ALL);
     }
 
     /**