+++ protected/humhub/modules/dashboard/messages/ja/views_dashboard_index_guest.php
@@ -1,4 +1,4 @@
 <?php
 return array (
-  '<b>No public contents to display found!</b>' => '',
+  '<b>No public contents to display found!</b>' => 'お見せできるようなコンテンツはまだないようですが、気をおとさず間をおいて再訪問してみましょう。',
 );