+++ protected/humhub/modules/like/messages/vi/controllers_LikeController.php
@@ -1,4 +1,4 @@
 <?php
 return array (
-  '<strong>Users</strong> who like this' => 'Những <strong>Users</strong> like điều này',
+  '<strong>Users</strong> who like this' => 'Những <strong>Người</strong> like điều này',
 );