+++ protected/humhub/modules/user/messages/vi/invite.php
@@ -1,8 +1,8 @@
 <?php
 return array (
   '<strong>Invite</strong> new people' => '<strong>Mời</strong> người mới',
-  'Email address(es)' => 'Địa chỉ Email (một hoặc nhiều)',
-  'Please add the email addresses of people you want to invite below.' => 'Vui lòng thêm các địa chỉ email của người bạn muốn thêm',
+  'Email address(es)' => '(Những) Địa chỉ Email',
+  'Please add the email addresses of people you want to invite below.' => 'Vui lòng thêm những địa chỉ email của người bạn muốn thêm',
   'Send invite' => 'Gửi lời mời',
   'Separate multiple email addresses by comma.' => 'Nhập địa chỉ email cách nhau bởi dấu phảy',
   'You\'ve been invited to join %appName%' => 'Bạn vừa được mời tham gia %appName%',