+++ protected/humhub/modules/user/messages/vi/views_profile_index.php
@@ -1,5 +1,5 @@
 <?php
 return array (
-  '<b>This profile stream is still empty!</b>' => '<b>Dòng thời gian này này vẫn chưa có nội dung!</b>',
+  '<b>This profile stream is still empty!</b>' => '<b>Dòng thời gian trên trang cá nhân này này vẫn chưa có nội dung!</b>',
   '<b>Your profile stream is still empty</b><br>Get started and post something...' => '<b>Trang cá nhân của bạn còn trống!</b><br>Hãy bắt đầu viết điều gì đó.',
 );