+++ protected/humhub/modules/user/messages/vi/widgets_BasePickerField.php
@@ -1,11 +1,11 @@
 <?php
 return array (
   'Add more...' => 'Thêm tiếp...',
-  'An unexpected error occurred while loading the result.' => 'Có lỗi không mong muốn xảy ra khi tải kết quả.',
+  'An unexpected error occurred while loading the result.' => 'Có lỗi không mong muốn xảy ra khi tải trang kết quả.',
   'Load more' => 'Tải thêm',
   'Please enter at least {n} character' => 'Vui lòng nhạp ít nhất {n} ký tự',
-  'Select {n,plural,=1{item} other{items}}' => 'Chọn {n,plural,=1{item} khác {items}}',
-  'This field only allows a maximum of {n,plural,=1{# item} other{# items}}.' => 'Trường này chỉ cho phép tối đa {n,plural,=1{#item} khác{# items}}.',
+  'Select {n,plural,=1{item} other{items}}' => 'Chọn {n,plural,=1{item} other{items}}',
+  'This field only allows a maximum of {n,plural,=1{# item} other{# items}}.' => 'Trường này chỉ cho phép tối đa {n,plural,=1{# item} other{# items}}.',
   'You reached the maximum number of allowed charachters ({n}).' => 'Bạn đã đạt tối đa số ký tự cho phép ({n}).',
   'Your search returned no matches.' => 'Tìm kiếm của bạn không trả về kết quả phù hợp nào.',
 );