+++ protected/humhub/modules/user/messages/vi/widgets_UserPickerField.php
@@ -2,6 +2,6 @@
 return array (
   'Add more...' => 'Thêm...',
   'No users found for the given query.' => 'Không tìm thấy người dùng theo tiêu chí tìm kiếm',
-  'Select {n,plural,=1{user} other{users}}' => 'Chọn {n,plural,=1{user} khác {users}}',
-  'This field only allows a maximum of {n,plural,=1{# user} other{# users}}.' => 'Trường này chỉ cho phép tối đa {n,plural,=1{# user} khác{# user}}.',
+  'Select {n,plural,=1{user} other{users}}' => 'Chọn {n,plural,=1{user} other{users}}',
+  'This field only allows a maximum of {n,plural,=1{# user} other{# users}}.' => 'Trường này chỉ cho phép tối đa {n,plural,=1{# user} other{# users}}.',
 );