+++ protected/humhub/docs/CHANGELOG.md
@@ -23,13 +23,17 @@ Important note for LDAP users: There is a new setting "ID Attribute" which shoul
 - Fix: trigger richtext `clear` when submitting comment.
 - Fix: missing return in `FileContent::beforeValidate`
 - Fix: Mentioning search with `-` not working
-- Fix #2730 Mentioning search with `-` not working
-- Fix File search with suffix not working
+- Fix #2730: Mentioning search with `-` not working
+- Fix: File search with suffix not working
+- Enh: Added SearchAttributesEvent to improve content addon indexing (comment/file)
+- Fix: Do not automatically force modal close on stream edit
+- Enh: Added DurationPickerWidget
 - Enh: Allow `ContentActiveRecord($contentContainer, $config)` initialization
 - Fix: `WallEntry::addControl` with simple array options
 - Enh: Added `$scheme` Argument to `DownloadFileHandler::getUrl()`
 - Fix: Clear UserModule cache after save/delete
 - Fix: Prevent Integrity check failures.
+
 
 1.2.2  (August 2, 2017)
 --------------------------------